requirements.txt
torch
transformers
gradio
Pillow
peft
datasets
trl
accelerate
sentence-transformers
faiss-cpu
numpy